Alertmanager: Support uploading Grafana Alertmanager Configuration and State
Grafana's Alertmanager and Mimir Alertmanager are configured differently but have the same set of upstream components running except receivers.

We'd like to enable Grafana to use the Mimir Alertmanager as a backend when Grafana is run with certain configuration so that Grafana can stop leveraging its internal Alertmanager.

One of the first steps in this direction is to allow the Mimir Alertmanager to store two things:

1. Grafana's Alertmanager Configuration
2. Grafana's Alertmanager State (Notification log and Silences)

This PR, setups up two sets of APIs to allow the Mimir the Mimir Alertmanager to support Grafana in this migration path. One for CRUDing the configuration and one for CRUDing the set of States.

Although these APIs are per tenant, they are not meant to be called by tenants and are experimental.