distributor, compactor, store-gateway: update ring heartbeat defaults
These are values which help reduce the pressure on a KV store or reduce the CPU spent by memberlist in passing messages.

THe tradeoff is that abrupt shutdowns/crashes of components will take longer to detect by peers. We've been running with these values at Grafana Labs for some time and haven't seen problems.

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>